requirements.txt
faster-whisper==1.0.1
deepfilternet==0.5.6
python-telegram-bot==13.15
python-ffmpeg==2.0.10